requirements.txt
langchain
langchain-groq
langchain-community
chromadb
pypdf2
streamlit
python-dotenv
sentence-transformers